removed start screen activity, it was pointless. fixed views and animations from splash screen to main activity. added timeout for fetching movies text. adjusted the app bar title to display once the fetching screen fades out
