+++ app/src/main/res/layout/activity_main.xml
@@ -2,7 +2,6 @@
 <android.support.constraint.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
-    xmlns:dots="http://schemas.android.com/apk/res-auto"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     tools:context=".MainActivity">
@@ -30,13 +29,13 @@
         android:layout_width="0dp"
         android:layout_height="match_parent"
         android:layout_marginBottom="8dp"
-        android:visibility="gone"
         app:layout_constraintBottom_toBottomOf="@+id/swipe_refresh"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent" />
 
     <include
         android:id="@+id/no_network"
+        android:visibility="gone"
         layout="@layout/no_network_view" />
 
     <include layout="@layout/no_results_view"
