--- app/src/main/res/layout/activity_start_screen.xml
@@ -1,76 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<android.support.constraint.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    xmlns:dots="http://schemas.android.com/apk/res-auto"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:background="@drawable/main_background"
-    tools:context=".StartScreenActivity">
-
-    <include
-        layout="@layout/app_header"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_marginTop="56dp"
-        app:layout_constraintTop_toTopOf="parent" />
-
-    <ImageView
-        android:id="@+id/movie_camera"
-        android:layout_width="@dimen/camera_image_width"
-        android:layout_height="@dimen/camera_image_height"
-        android:contentDescription="@string/movie_camera_image_content_description"
-        android:scaleX="1.25"
-        android:scaleY="1.25"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent"
-        app:srcCompat="@drawable/movie_camera" />
-
-    <TextView
-        android:id="@+id/app_tag_line"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginBottom="@dimen/margins_8"
-        android:layout_marginEnd="@dimen/margins_8"
-        android:layout_marginStart="@dimen/margins_8"
-        android:fontFamily="@font/futura_pt_book"
-        android:text="@string/app_description_title_screen"
-        android:textAllCaps="true"
-        android:textColor="@color/textStartCredits"
-        android:textSize="@dimen/text08"
-        android:textStyle="bold"
-        app:layout_constraintBottom_toTopOf="@+id/credits_bar"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent" />
-
-    <View
-        android:id="@+id/credits_bar"
-        android:layout_width="0dp"
-        android:layout_height="@dimen/title_screen_credits_bar_height"
-        android:layout_marginBottom="@dimen/margins_8"
-        android:layout_marginEnd="@dimen/margin_28"
-        android:layout_marginStart="@dimen/margin_28"
-        android:background="@color/old_rose"
-        app:layout_constraintBottom_toTopOf="@+id/credits_text"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent" />
-
-    <TextView
-        android:id="@+id/credits_text"
-        android:layout_width="0dp"
-        android:layout_height="wrap_content"
-        android:layout_marginBottom="@dimen/margin_28"
-        android:layout_marginEnd="@dimen/margins_24"
-        android:layout_marginStart="@dimen/margins_24"
-        android:fontFamily="@font/credit_block"
-        android:text="charles rowland presents “movie night” in association with udacity as part of the android develoer nanodegree an app by charles rowland art direction lillian rowland and charles rowland distractions by ethan rowland emotional support daisey the dog development music a lot of heavy metal and always remember thanos did nothing wrong"
-        android:textAlignment="center"
-        android:textColor="@color/textStartCredits"
-        android:textSize="@dimen/text18"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent" />
-
-</android.support.constraint.ConstraintLayout>
