+++ app/src/main/res/layout/fetching_movies_view.xml
@@ -12,7 +12,6 @@
         layout="@layout/app_header"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
-        android:layout_marginTop="0dp"
         app:layout_constraintTop_toTopOf="parent" />
 
     <ImageView
@@ -39,8 +38,7 @@
         android:text="@string/app_description_title_screen"
         android:textAllCaps="true"
         android:textColor="@color/textStartCredits"
-        android:textSize="@dimen/text08"
-        android:textStyle="bold"
+        android:textSize="@dimen/text10"
         app:layout_constraintBottom_toTopOf="@+id/credits_bar"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent" />
@@ -50,8 +48,8 @@
         android:layout_width="0dp"
         android:layout_height="@dimen/title_screen_credits_bar_height"
         android:layout_marginBottom="@dimen/margins_8"
-        android:layout_marginEnd="@dimen/margin_28"
-        android:layout_marginStart="@dimen/margin_28"
+        android:layout_marginEnd="@dimen/margins_14"
+        android:layout_marginStart="@dimen/margins_14"
         android:background="@color/old_rose"
         app:layout_constraintBottom_toTopOf="@+id/fetching_text"
         app:layout_constraintEnd_toEndOf="parent"
@@ -62,15 +60,15 @@
         style="@style/fetchingMovies"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
-        android:layout_marginStart="@dimen/margins_24"
-        android:layout_marginEnd="@dimen/margins_24"
-        android:layout_marginBottom="46dp"
+        android:layout_marginEnd="@dimen/margins_20"
+        android:layout_marginStart="@dimen/margins_20"
+        android:layout_marginBottom="36dp"
         android:fontFamily="@font/futura_pt_book"
         android:gravity="center_horizontal"
         android:text="@string/getting_movies"
         android:textColor="@color/textStartCredits"
         android:textAllCaps="true"
-        android:textSize="14sp"
+        android:textSize="@dimen/text14"
         app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent" />
@@ -81,12 +79,12 @@
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:layout_marginStart="@dimen/margins_8"
-        android:layout_marginTop="@dimen/margins_8"
+        android:layout_marginTop="@dimen/margins_12"
         android:layout_marginEnd="@dimen/margins_8"
         android:textSize="@dimen/text45"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/fetching_text"
+        app:layout_constraintTop_toTopOf="@+id/fetching_text"
         dots:autoplay="true"
         dots:period="1000" />
