+++ app/src/main/res/values/arrays.xml
@@ -13,7 +13,7 @@
     <string-array name="title_screen_credits">
         <item>charles rowland</item>
         <item>presents</item>
-        <item>“movie night”</item>
+        <item>\"movie night\"</item>
         <item>in association with</item>
         <item>udacity as part of the</item>
         <item>android develoer nanodegree</item>
@@ -31,5 +31,31 @@
         <item>a lot of heavy metal</item>
         <item>and always remember</item>
         <item>thanos did nothing wrong</item>
+    </string-array>
+
+    <string-array name="title_screen_credits_big">
+        <item>charles rowland</item>
+        <item>“movie night”</item>
+        <item>udacity as part of the</item>
+        <item>an app by</item>
+        <item>art direction</item>
+        <item>and</item>
+        <item>distractions by</item>
+        <item>emotional support</item>
+        <item>development music</item>
+        <item>and always remember</item>
+    </string-array>
+
+    <string-array name="title_screen_credits_normal">
+        <item>charles rowland</item>
+        <item>“movie night”</item>
+        <item>udacity as part of the</item>
+        <item>an app by</item>
+        <item>art direction</item>
+        <item>and</item>
+        <item>distractions by</item>
+        <item>emotional support</item>
+        <item>development music</item>
+        <item>and always remember</item>
     </string-array>
 </resources>
