+++ app/src/main/res/values/strings.xml
@@ -47,4 +47,6 @@
     <string name="movie_camera_image_content_description">movie camera image</string>
     <string name="app_description_title_screen">Popular and top rated movie information in one convenient place</string>
 
+    <string name="title_screen_credits_string">charles rowland presents “movie night” in association with udacity as part of the android develoer nanodegree an app by charles rowland art direction lillian rowland and charles rowland distractions by ethan rowland emotional support daisey the dog development music a lot of heavy metal and always remember thanos did nothing wrong</string>
+
 </resources>
